% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveBetaTree.R
\name{build_adaptive_histogram}
\alias{build_adaptive_histogram}
\title{Build an Adaptive Beta-Tree Histogram}
\usage{
build_adaptive_histogram(
  X,
  alpha = 0.1,
  thresh_marginal = alpha/ncol(X),
  thresh_interaction = stats::qgamma(shape = ncol(X) - 1, rate = 1, p = 1 - alpha),
  method = "weighted_bonferroni"
)
}
\arguments{
\item{X}{A numeric matrix dimension \eqn{n \times d} containing observations falling inside the current
region. Each row corresponds to one observation.}

\item{alpha}{Significance level, default is \code{alpha = 0.1}.
The Beta-tree histogram provides simultaneous confidence intervals for average densities
in each histogram region with coverage probability \eqn{1-\alpha}.}

\item{thresh_marginal}{A numeric value between \eqn{(0, 1)}. The significance
threshold for the marginal Anderson-Darling uniformity test. Default value is \code{alpha / d}.}

\item{thresh_interaction}{A positive numeric value. The threshold on the
combined interaction score \eqn{\sum_{j\neq i} -\log p_{ij}} for coordinate
\eqn{i}, where \eqn{p_{ij}} is the p-value of testing independence between \eqn{X_i} and \eqn{X_j}.
Default value is \eqn{(1-\alpha)} quantile of a Gamma variable with shape \eqn{d-1} and rate 1.}

\item{method}{A character string specifying whether to use \code{method = "bonferroni"} or \code{method = "weighted_bonferroni"} method to
adjust for multiple hypothesis testing when computing significance levels.}
}
\value{
A named list with three components:
\describe{
\item{\code{split_dir}}{An integer vector recording the splitting
coordinate chosen at each step. The first element is always \code{NA}. }
\item{\code{hist}}{ A numeric matrix of dimension \eqn{m \times (2d + 5)}.
Each row represents one region in the Beta-tree histogram.
The first \eqn{d} columns are the lower bounds for the region; the next \eqn{d} columns are the upper bounds.
The final 5 columns give the empirical density, lower and upper confidence bounds of average density, ,
upper confidence bound, number of observations inside, and tree depth of each region. See also \code{\link{BuildHist}}.}
\item{\code{lower}}{A numeric vector of length \eqn{d} corresponding to the
the minimum observation in each coordinate.}
\item{\code{upper}}{A numeric vector of length \eqn{d} corresponding to the
the maximum observation in each coordinate.}
}
}
\description{
Constructs an Beta-tree histogram there partition coordinates are chosen adaptively
based on tests of marginal uniformity and pairwise independence.
}
\details{
The adaptive Beta-tree histogram uses the same algorithm as the Beta-tree histogram (see \code{\link{BuildHist}} function
and package Vignettes \code{beta_trees}) with two differences:

First, the adaptive Beta-tree automatically computes a bounded histogram by placing the first splits at the
minimum and maximum of each coordinates.

Second, the adaptive Beta-tree uses \code{\link{get_partition_dim}} to select partition dimension
at each step. The same set of thresholds \code{thresh_marginal} and \code{thresh_interaction} are used at every step.
}
\examples{
set.seed(42)

# Bivariate standard normal data
X <- matrix(rnorm(1000), nrow = 500, ncol = 2)

adaptive_beta_tree <- build_adaptive_histogram(
  X                 = X,
  alpha             = 0.1,
  method            = "weighted_bonferroni"
)

# Number of histogram regions
nrow(adaptive_beta_tree$hist)

# Splitting directions
table(adaptive_beta_tree$split_dir)

}
\seealso{
\code{\link{get_partition_dim}} adaptively selectx partition coordinates.
\code{\link{BuildHist}} builds a Beta-tree histogram.
}
