% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildKDTree.R
\name{BuildKDTree}
\alias{BuildKDTree}
\title{Build a k-d tree}
\usage{
BuildKDTree(X, bounded = F, ...)
}
\arguments{
\item{X}{data matrix with n rows and d columns}

\item{bounded}{if bounded = TRUE, then creates an initial bounding box according to user input parameters:
\describe{
\item{option}{Define the bounding box by the number of obs. (option = ndat) or quantiles (option = qt) of observations included.}
\item{q}{number/quantile of points to be excluded in each end of a dimension. For example, if data is two-dimensional, option = "ndat" and qt = 50, then exclude 50 obs. in the lower and upper end of x-coordinate, then for the observations within, exclude 50 obs. in the lower and upper end of y-coordinate.  }
}}

\item{...}{Additional parameters if bounded = TRUE}
}
\value{
\code{build_tree} returns two values.
\describe{
\item{kdtree} the k-d tree built from root node.
\item{nd} The number of bounded regions at every level, starting at level 1.
}
}
\description{
Grow a kd-tree by iteratively partitioning along the sample median of each coordinate. Stop partitioning when the number of observatins in a node is less than \eqn{4\log(n)}
}
\details{
\code{BuildKDTree} constructs a kd-tree from data by iteratively partitioning along the sample median each coordinate.
Each node in the kd-tree represents a region, and it consists of the following information:
\describe{
\item{depth}{level of the region. The depth of root node is 1.}
\item{ndat}{number of samples in this region.}
\item{low, up}{lower and upper bound on region, if it exists.}
\item{lower, upper}{lower and upper confidence bound for average density.}
\item{bounded}{whether the region is bounded. A region is bounded if all of (\code{low}, \code{up}) exist.}
\item{leaf}{whether the node is leaf or not. A node is a leaf if the number of observations is less than \eqn{4 \log n}, where \eqn{n} is the total number of samples. A node is partitioned if it is not a leaf. }
\item{leftchild, rightchild}{pointers to the left and right child of a node. \code{leftchild} points to the node representing the region of observations \emph{below} median in partition dimension, and \code{rightchild} points to the node representing the region of observations \emph{above} the median.}
}
}
\examples{
# build a k-d tree for standard normal samples
X <- matrix(rnorm(1000), ncol = 2)
kd <- BuildKDTree(X)
}
