% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_adjacencyMat.R
\name{compute_adjacency_mat}
\alias{compute_adjacency_mat}
\title{Adjacency matrix of regions in the Beta Tree}
\usage{
compute_adjacency_mat(hist, d)
}
\arguments{
\item{hist}{A matrix. Each row represents one region in the histogram.
If the histogram is constructed for \code{d} dimensional data, then the first \code{d} columns contain the lower bound of each dimension,
and the next \code{d} columns contain the upper bound. \code{hist} can be the output of \link[BetaTrees]{BuildHist} function.}

\item{d}{A numeric value of the data dimension.}
}
\value{
A matrix of size N*N, where N is the number of regions in the histogram, i.e. number of rows in the histogram.
}
\description{
Computes the adjacency matrix of all of the regions in a Beta Tree histogram
}
\details{
Two regions are neighbors if and only if the interval (defined by the boundary) in each dimension intersects.
For example, let the two regions \eqn{X} defined by \eqn{[X_{i,\mathrm{low}}], X_{i,\mathrm{up}}]}
and \eqn{Y} defined by \eqn{[Y_{i,\mathrm{low}}], Y_{i,\mathrm{up}}]} (for \eqn{i=1,\ldots, d}), where \eqn{[X_{i,\mathrm{low}}]} and \eqn{X_{i,\mathrm{up}}]}
are the lower and upper bounds in dimension \eqn{i} for the region \eqn{X}. Then, \eqn{X} and \eqn{Y} are neighbors if and only if
\eqn{[X_{i,\mathrm{low}}], X_{i,\mathrm{up}}] \cap [Y_{i,\mathrm{low}}], Y_{i,\mathrm{up}}] \neq \emptyset} for every \eqn{i=1,\ldots, d}.
}
\examples{
X <-  matrix(rnorm(2000, 0, 1), ncol = 2)
rect <- BuildHist(X)
adj <- compute_adjacency_mat(hist = rect, d = 2)
}
