% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_mode.R
\name{connected_to_modes}
\alias{connected_to_modes}
\title{Is a Region a Distinct Mode? (Approximate Algorithm)}
\usage{
connected_to_modes(candidate_mode, curr_mode, g, cluster, ci, L, B)
}
\arguments{
\item{candidate_mode}{A positive integer giving the row index of
the candidate region.}

\item{curr_mode}{An integer vector of row indices identifying the
currently identified mode regions in the histogram.}

\item{g}{An \code{igraph} undirected graph object representing the
adjacency graph of the histogram regions.}

\item{cluster}{A list of length \code{length(curr_mode)}, where the
\eqn{k}-th element is a vector indicating regions in the
neighborhood of mode \eqn{k}, as returned by
\code{\link{find_neighbors}}.}

\item{ci}{A numeric matrix with two columns giving the lower
and upper confidence bounds for the average density of each histogram region.}

\item{L}{A positive integer. The length of random path.}

\item{B}{A positive integer. The number of independent random paths
used to assess whether a candidate region is connected to any of the current modes.}
}
\value{
A named list with one element:
\describe{
\item{\code{connected}}{A character string: \code{"connected"} if
the candidate is \emph{not} a distinct mode and \code{"unconnected"} otherwise.}
}
}
\description{
Tests whether a histogram region is connected to any of the
current modes using Monte Carlo algorithm.
}
\details{
Generate \code{B} random paths of length \code{L} starting from
\code{candidate_mode} and evaluate (1) whether it connects the \code{candidate_mode}
to any \code{curr_mode}; if so, (2) whether there exists a region along the path whose upper
confidence bound is below the lower confidence bounds of both the \code{candidate_mode} and
the \code{curr_mode}. If such a region does not exist, then the \code{candidate_mode} \emph{cannot} be a distinct mode.
}
\seealso{
\code{\link{FindModesApproximate}} calls this function to test whether a candidate region is a distince mode.
}
