% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{FindModes}
\alias{FindModes}
\title{Compute modes in the Beta Tree histogram}
\usage{
FindModes(hist, d, cutoff = 6)
}
\arguments{
\item{hist}{A matrix. Each row represents one region in the histogram.
If the histogram is constructed for \code{d} dimensional data, then the first \code{d} columns contain the lower bound of each dimension,
and the next \code{d} columns contain the upper bound. The \code{2d + 1} column contains the empirical density.
\code{hist} can be the output of \link[BetaTrees]{BuildHist} function.}

\item{d}{A numeric value of the data dimension.}

\item{cutoff}{Maximum path length. By default, we only evaluate paths with length at most \code{cutoff = 6}.}
}
\value{
\describe{
\item{mode}{A vector of candidate modes.}
\item{hist}{A matrix of the input Beta Tree histogram.}
\item{g}{A graph object constructed using the adjacency matrix of the regions in the histogram.}
}
}
\description{
Identifies distinct modes in the histogram
}
\details{
Two regions \code{i} and \code{j} are distinct modes if along at least one path connecting them,
there exists one region whose upper confidence limit is below the lower confidence limit
of both \code{i} and \code{j}.
This function iterates through all of the regions in the Beta Tree histogram according to the
descending order of empirical density and marks a region as a distinct mode if it is a distinct mode
compared to every current modes.
}
\examples{
# A mixture of two Gaussians
X <-  matrix(rnorm(4000, 0, 1), ncol = 2)
X[1:1000, 1] <- X[1:1000, 1] + 4
rect <- BuildHist(X) # construct a Beta Tree histogram
modes <- FindModes(hist = rect, d = 2)$mode
}
