% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_mode.R
\name{FindModesApproximate}
\alias{FindModesApproximate}
\title{Identify Modes in a Beta-tree Histogram via Monte Carlo approximation}
\usage{
FindModesApproximate(hist, d, L, B, verbose = FALSE)
}
\arguments{
\item{hist}{A numeric matrix of dimension \eqn{m\times (2d+5)}.
Each row represents one region in the histogram.
If the histogram is constructed for \code{d} dimensional data,
then the \eqn{1:d} columns contain the lower bounds for each coordinate,
columns \eqn{(d+1):(2d)} contain the upper bounds.
The \eqn{2d + 1} column contains the empirical density.
The \eqn{(2d+2):(2d+3)} columns contain the lower and upper confidence bounds of the empirical density.
\code{hist} can be the output of \link[BetaTree]{BuildHist} or \link[BetaTree]{build_adaptive_histogram} function.}

\item{d}{A positive integer corresponding to the data dimension.}

\item{L}{A positive integer. The length of random path.}

\item{B}{A positive integer. The number of independent random paths
used in \code{\link{connected_to_modes}} to assess whether a candidate region
is connected to any of the current modes.}

\item{verbose}{Prints progress  if \code{TRUE}.}
}
\value{
A named list with three components:
\describe{
\item{\code{mode}}{An integer vector of row indices of \code{hist}
identifying the detected modes.}
\item{\code{hist}}{The input \code{hist} matrix.}
\item{\code{cluster}}{A list of length \code{length(mode)}. The
\eqn{k}-th element is an integer vector of row indices of histogram regions
that belonging to the neighborhood of mode \eqn{k}. That is, there is a path from this region to the mode
that the upper confidence bound of every region along the path (including the region itself) is higher than the lower confidence bound of the mode. }
}
}
\description{
Implements an approximate mode-hunting algorithm by using randomly sampled paths
to estimate whether a region is a distinct mode.
}
\details{
Evaluate every region in order of decreasing density whether it is a distinct mode compared to current modes.
In the exact algorithm \link{FindModes}, a region \eqn{R} is considered a distinct mode from a current model \eqn{M} if
if along every path connecting \eqn{R} and \eqn{M}, there exists at least one region whose upper confidence bound of density
is below the lower confidence bounds of both \eqn{R} and \eqn{M}. The approximate algorithm here evaluates \code{B} random paths of length \code{L}
instead of evaluating every path.
}
\examples{
\dontrun{
# mixture of two Gaussian distribution
X <- rbind(matrix(rnorm(500, mean = -2), ncol = 2),
           matrix(rnorm(500, mean =  2), ncol = 2))
beta_tree_hist <- BuildHist(X)

# find modes using approximate algorithm
modes <- FindModesApproximate(hist  = beta_tree_hist,
                              d     = 2,
                              L     = 50,
                              B     = 1000,
                              verbose = F)

# number of detected modes
length(modes$mode)
}

}
