% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multivariate_hist.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot a multivariate histogram in two dimensions}
\usage{
plot_histogram(plot_data, show_data = TRUE)
}
\arguments{
\item{plot_data}{A named list as returned by
\code{\link{plot_histogram_data}}.}

\item{show_data}{Logical. If \code{TRUE} (default), shows the data
points in \code{plot_data$data}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Visualize an adaptive beta-tree histogram in two-dimensions marginally or conditional on inside a hyperrectangle.
}
\details{
\code{\link{plot_histogram_data}} computes the probability content in each region,
\code{plot_histogram} further computes the conditional distribution after dividing by the total probability content.
}
\examples{

\dontrun{
n <- 10000
p <- 4
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
adaptive_beta_tree <- build_adaptive_histogram(X, alpha = 0.1)

plot_dat <- plot_histogram_data(
  X          = X,
  hist       = adaptive_beta_tree$hist,
  plot_coord = c(1, 2),
  nint       = 20,
  show_data  = TRUE,
  ndat       = 500
)

plot_histogram(plot_dat)
}

}
\seealso{
\code{\link{plot_histogram_data}} for computing the input \code{plot_data},
\code{\link{build_adaptive_histogram}} for producing the histogram.
}
