% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectNodes.R
\name{SelectNodes}
\alias{SelectNodes}
\title{Select nodes in the histogram}
\usage{
SelectNodes(node, B, ahat, n)
}
\arguments{
\item{node}{A k-d tree output from \code{SetBounds}.}

\item{B}{A matrix containing nodes that have been selected.}

\item{ahat}{A vector of significance levels at each depth, starting at depth = 0. Output of the funtion \code{ConfLevel}}

\item{n}{Total number of observations.}
}
\value{
A matrix of selected regions within and including the input node.
Each row represents one selected region and has \eqn{2d + 5} columns, containing the lower and upper bounds of the region,
the empirical density, lower and upper confidence bounds of the average density in this region,
number of obs. inside and the depth of the node.
}
\description{
Extract maximal (w.r.t. inclusion) nodes that are
bounded and whose empirical density is within confidence bounds.
}
\details{
A bounded node is selected if the empirical density is inside the confidence bounds of all of its children.
The empirical density is defined as
\deqn{
h = \frac{n+1}{N |R|},
}
where \eqn{N} is the total number of obs., \eqn{n} is the number of obs. in the region,
and \eqn{|R|} is the volume of the region.
If the node is not selected and is not a leaf, then move on to the left and right children of the node.
}
