% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildHist.R
\name{BuildHist}
\alias{BuildHist}
\title{Build a Beta Tree}
\usage{
BuildHist(
  X,
  alpha = 0.1,
  method = "weighted_bonferroni",
  bounded = F,
  plot = F,
  ...
)
}
\arguments{
\item{X}{a data matrix of size n*p.}

\item{alpha}{confidence level, default level is \code{alpha = 0.1}.}

\item{method}{use \code{method = "bonferroni"} or \code{method = "weighted_bonferroni"}(default) to adjust for multiple hypothesis testing.}

\item{bounded}{if bounded = TRUE, then creates an initial bounding box according to user input parameters:
\describe{
\item{option}{Define the bounding box by the number of obs. (option = ndat) or quantiles (option = qt) of observations included.}
\item{q}{number/quantile of points to be excluded in each end of a dimension. For example, if data is two-dimensional, option = "ndat" and qt = 50, then exclude 50 obs. in the lower and upper end of x-coordinate, then for the observations within, exclude 50 obs. in the lower and upper end of y-coordinate.  }
}}

\item{plot}{if \code{TRUE} (default), plot the histogram if the data is two-dimensional.}

\item{...}{Additional parameter if \code{bounded = T}.}
}
\value{
A matrix of rectangles in the histogram of obs. in the input node.
A region in the histogram is represented by a row, which contains the following columns: the lower bounds, upper bounds, empirical density,
lower confidence bounds, upper confidence bounds of this node, number of obs. and the depth of the node.
}
\description{
Compute a Beta Tree histogram for multivariate data at given confidence level.
}
\details{
Construct the Beta Tree histogram by iteratively partitioning sample space along the sample median.
Then, construct \emph{simultaneous} confidence interval (at level \eqn{\alpha}) for all of the regions using Bonferroni
or weighted Bonferroni adjustment. Finally, select the largest bounded regions (w.r.t. inclusion) that that pass a goodness of fit (GOF) test, i.e.,
the empirical density lies within the confidence interval of both a node and both of its children.
The function can plot the histogram if the data is two-dimensional.
}
\section{Multiple testing correction}{


If Bonferroni correction is selected, then the confidence level at each region is defined as
}

\examples{
X <-  matrix(rnorm(2000, 0, 1), ncol = 2)
rect <- BuildHist(X)
# Plot the histogram
rect <- BuildHist(X, plot = TRUE)
# set confidence level and multiple hypothesis testing correction
rect <- BuildHist(X, alpha = 0.05, method = "weighted_bonferroni")
}
