% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfLevel.R
\name{ConfLevel}
\alias{ConfLevel}
\title{Set confidence level to each depth}
\usage{
ConfLevel(nd, alpha, method)
}
\arguments{
\item{nd}{A vector. Number of bounded regions at every level.}

\item{alpha}{Numeric. confidence level.}

\item{method}{Use \code{bonferroni} or \code{weighted_bonferroni} method to adjust for multiple hypothesis testing.}
}
\value{
A vector of confidence level at every depth (starting at depth = 1)
}
\description{
Computes the confidence level at every level of the tree.
}
\details{
If \code{method = Bonferroni}, the confidence level at depth d is
\deqn{
\hat{\alpha}_d = \frac{\alpha}{(d_\max - d_\min + 1) * n_d}, D\geq D_\min,
}
where \eqn{d_\max} is the maximum depth of the tree, \eqn{d_\min} is the smallest depth where there are bounded regions, and \eqn{n_d} is number of bounded regions at depth \eqn{d}.

If \code{method = weighted_bonferroni}, the confidence level at depth d is
\deqn{
\hat{\alpha}_d = \frac{\alpha}{(d_\max - d + 2) n_d \sum_{B=2}^{d_\max - d_\min + 2}1/B},D\geq D_\min,
}
that is, the leaves get weights alpha/2 etc.
}
