% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{FindModes}
\alias{FindModes}
\title{Identifies distinct modes in Beta-trees histogram}
\usage{
FindModes(hist, d, cutoff = 6)
}
\arguments{
\item{hist}{A matrix. Each row represents one region in the histogram.
If the histogram is constructed for \code{d} dimensional data, then the first \code{d} columns contain the lower bounds for each coordinate,
and the next \code{d} columns contain the upper bounds. The \code{2d + 1} column contains the empirical density.
\code{hist} can be the output of \link[BetaTree]{BuildHist} or \link[BetaTree]{SelectNodes} function.}

\item{d}{A numeric value of the data dimension.}

\item{cutoff}{Maximum path length. By default, we only evaluate paths with length at most \code{cutoff = 6}.}
}
\value{
\describe{
\item{mode}{A vector of candidate modes.}
\item{hist}{The input Beta-tree histogram.}
\item{g}{A graph object constructed using the adjacency matrix of the regions in the histogram.}
}
}
\description{
Compute modes in the histogram.
}
\details{
Two regions \code{i} and \code{j} are distinct modes if along every path connecting them,
there exists one region whose upper confidence limit is below the lower confidence limit
of both \code{i} and \code{j}.
This function iterates through all the regions in the Beta-tree histogram in
descending order of empirical density and marks a region as a mode if it is a distinct mode
from all currently discovered distinct modes.
}
\examples{
# A mixture of two Gaussians
X <-  matrix(rnorm(6000, 0, 1), ncol = 2)
X[1:1500, 1] <- X[1:1500, 1] + 4
rect <- BuildHist(X) # construct a Beta-tree histogram
modes <- FindModes(hist = rect, d = 2)$mode
rect[modes, ] # show information of the modes
}
