% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveBetaTree.R
\name{get_partition_dim}
\alias{get_partition_dim}
\title{Choose partition dimension}
\usage{
get_partition_dim(x, thresh_marginal, thresh_interaction, lower, upper)
}
\arguments{
\item{x}{A numeric matrix dimension \eqn{n \times d} containing observations falling inside the current
region. Each row corresponds to one observation.}

\item{thresh_marginal}{A numeric value between \eqn{(0, 1)}. The significance
threshold for the marginal Anderson-Darling uniformity test.}

\item{thresh_interaction}{A positive numeric value. The threshold on the
combined interaction score \eqn{\sum_{j\neq i} -\log p_{ij}} for coordinate
\eqn{i}, where \eqn{p_{ij}} is the p-value of testing independence between \eqn{X_i} and \eqn{X_j}}

\item{lower}{A numeric vector of length \eqn{d} specifying the lower
bounds of the current region.}

\item{upper}{A numeric vector of length \eqn{d} specifying the upper
bounds of the current region.}
}
\value{
A named list with three elements:
\describe{
\item{\code{below_marginal}}{A vector of integers corresponding to coordinates
whose marginal Anderson-Darling p-value is below
\code{thresh_marginal}, \code{NA} if no rejection.}
\item{\code{below_interaction}}{A vector of integers corresponding to coordinates
whose combined interaction score \eqn{p_i} exceeds
\code{thresh_interaction}, \code{NA} if no rejection.}
\item{\code{p}}{Selected partition coordinate.}
}
}
\description{
Selects partition dimension by testing marginal uniformity and pairwise independence.
}
\details{
Adaptively choosing partition dimension by first testing marginal uniformity
using Anderson-Darling test for each coordinate \eqn{i \in \{1, \ldots, d\}}, i.e.
we test if \eqn{x_{1i}, \ldots, x_{ni}} are uniformly distributed within
\eqn{(\texttt{lower}_i, \texttt{upper}_i)}. If the minimum p-value is less than \code{thresh_marginal},
randomly sample one coordinate with p-value less than \code{thresh_marginal} as the partition dimension.

Otherwise, test pairwise independence between all \eqn{\binom{d}{2}} coordinate pairs \eqn{(i, j)} using a Fisher's exact test.
To test if \eqn{X_i} is independent of \eqn{X_j} in the region \eqn{R}, we divide \eqn{R} into four
quadrants along the middle of \eqn{i} and \eqn{j} coordinates and apply Fisher's exact test to counts of observations
in each quadrant. Compute a combined p-value for \eqn{i}th coordinate
as \eqn{p_i = \sum_{j \neq i} -\log p_{ij}} and randomly sample one coordinate with \eqn{p_i > }\code{thresh_interaction} as partition dimension.

If no coordinate is significant in either tests, randomly pick one coordinate to partition.
}
\examples{
set.seed(42)

# Uniform observations
x <- matrix(runif(300), nrow = 100, ncol = 3)
get_partition_dim(x,
                 thresh_marginal    = 0.05,
                 thresh_interaction = qgamma(p=1-0.05, shape=3-1, rate=1),
                 lower = rep(0, 3),
                 upper = rep(1, 3))

# Non-uniform observations
y <- matrix(runif(300), nrow = 100, ncol = 3)
y[, 1] <- rbeta(100, 3, 3)
get_partition_dim(y,
                 thresh_marginal    = 0.05,
                 thresh_interaction = qgamma(p=1-0.05, shape=3-1, rate=1),
                 lower = rep(0, 3),
                 upper = rep(1, 3))

}
\seealso{
\code{\link{build_adaptive_histogram}} for the function that calls
\code{get_partition_dim} to adaptively choose partition locations.
}
