% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHist.R
\name{PlotHist}
\alias{PlotHist}
\title{Plot a 2D Beta tree histogram}
\usage{
PlotHist(X, B, nsample = 10000, ...)
}
\arguments{
\item{X}{data matrix of size n by 2.}

\item{B}{matrix of rectangles returned by the function \link[BetaTree]{SelectNodes} or \link[BetaTree]{BuildHist}.}

\item{nsample}{maximum number of data points to display in the plot.}

\item{...}{additional parameter for the graph: \code{low}, \code{mid}, \code{high} specifies the colors in \link[ggplot2]{scale_fill_gradient2} function.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a Beta tree histogram.
}
\details{
\code{PlotHist} plots the histogram superimposed on the original dataset.
If the number of data points exceeds \code{nsample}, randomly select \code{nsample} obs. to show in the graph. Default value is \eqn{nsample=10^4}
}
\examples{
X <-  matrix(rnorm(2000, 0, 1), ncol = 2)
rect <- BuildHist(X, plot = FALSE)
PlotHist(X, rect)
}
