% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectNodes.R
\name{SelectNodes}
\alias{SelectNodes}
\title{Select nodes in the histogram}
\usage{
SelectNodes(node, B, ahat, n)
}
\arguments{
\item{node}{A k-d tree with confidence thresholds, output from \code{SetBounds}.}

\item{B}{A matrix.}

\item{ahat}{A vector of confidence levels at each depth, starting at depth = 1. Output of the funtion \code{ConfLevel}}

\item{n}{Total number of observations.}
}
\value{
A matrix of rectangles in the histogram of obs. in the input node.
A region in the histogram is represented by a row, which contains the lower and upper bounds, empirical density,
lower and upper confidence bounds of this node, number of obs. and the depth of the node.
}
\description{
Extract maximal (w.r.t. inclusion, i.e. tree order) nodes that are
bounded and whose empirical density is within confidence bounds.
}
\details{
A bounded node is selected if the empirical density is within the confidence bounds.
The empirical density is
\deqn{
h = \frac{n+1}{N |R|},
}
where \eqn{N} is the number of obs., and \eqn{n} is the number of obs. in the region,
and \eqn{|R|} is the volume of the region.
Otherwise, if the node is not a leaf, move on to the left and right children of the node.
}
